requirements.txt
absl-py==2.3.1
accelerate==1.12.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
attrdict==2.0.1
attrs==25.4.0
audioread==3.1.0
av==16.0.1
binpacking==1.5.2
bitsandbytes==0.49.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cpm-kernels==1.0.11
crcmod==1.7
cryptography==46.0.3
cycler==0.12.1
dacite==1.9.2
datasets==3.6.0
decorator==5.2.1
dill==0.3.8
distro==1.9.0
docker==7.1.0
einops==0.8.1
fastapi==0.127.0
ffmpy==1.0.0
filelock==3.20.1
flash_attn==2.8.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.3.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
gradio==5.50.0
gradio_client==1.14.0
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.12.0
jmespath==0.10.0
joblib==1.5.3
json_repair==0.54.3
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
modelscope==1.33.0
mpmath==1.3.0
ms_swift @ git+https://github.com/modelscope/ms-swift.git@bf53f1433358fd5c43e5913fd56f2c9ac1a7d53d
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
networkx==3.6.1
nltk==3.9.2
numba==0.63.1
numpy==2.3.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==2.14.0
orjson==3.11.5
oss2==2.19.1
packaging==25.0
pandas==2.3.3
peft==0.18.0
pillow==11.3.0
platformdirs==4.5.1
pooch==1.8.2
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pyarrow==22.0.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
qwen-omni-utils==0.0.8
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rouge==1.0.1
ruff==0.14.10
safehttpx==0.1.7
safetensors==0.7.0
sca_data @ git+https://github.com/riverfog7/sca_data_prep@4de394466b736a2f74db51b456365ad4deed7067
scikit-learn==1.8.0
scipy==1.16.3
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.48.0
setuptools==70.2.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
starlette==0.50.0
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.57.3
transformers-stream-generator==0.0.5
triton==3.1.0
trl==0.24.0
typer==0.20.1
typer-slim==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
wandb==0.23.1
websockets==15.0.1
Werkzeug==3.1.4
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
